+++ frontend/src/components/frontpage.vue
@@ -1,0 +1,14 @@
+<template>
+    <div class="what">
+    </div>
+</template>
+
+<script setup lang="ts">
+
+</script>
+
+<style scoped>
+.what{
+    background-image: url("/concept1.png")
+}
+</style>
