+++ frontend/src/main.ts
@@ -10,7 +10,7 @@ import Ripple from 'primevue/ripple';
 import App from './App.vue'
 import router from './router'
 import PrimeVue from 'primevue/config'
-import Aura from '@primevue/themes/aura'
+
 
 const app = createApp(App)
 
@@ -20,7 +20,8 @@ app.use(router)
 app.use(PrimeVue, {
   theme: {
     preset: Aura
-  }
+  },
+  ripple: true,
 })
 
 app.mount('#app')
